bug fixed

+++ app/src/main/java/edu/utcs/comicwiki/ui/creation/CreationViewModel.kt
@@ -1,14 +1,8 @@
 package edu.utcs.comicwiki.ui.creation
 
-import android.app.Activity
-import android.app.Application
-import android.widget.Toast
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
+import android.util.Log
+import androidx.lifecycle.*
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.firestore.FirebaseFirestore
 import edu.utcs.comicwiki.model.ComicNode
 
@@ -36,6 +30,7 @@ class CreationViewModel() : ViewModel() {
     fun observeCenterNode(): LiveData<ComicNode> {
         return centerNode
     }
+
     fun deleteCenterNode() {
         centerNode.value = ComicNode()
     }
@@ -50,10 +45,10 @@ class CreationViewModel() : ViewModel() {
         return relatedNodes
     }
 
-    fun addRelatedNode(comicNode: ComicNode) {
+    fun addRelatedNode(relatedNode: ComicNode) {
         val local = relatedNodes.value?.toMutableList()
         local?.let {
-            it.add(comicNode)
+            it.add(relatedNode)
             relatedNodes.value = it
         }
     }
@@ -87,12 +82,16 @@ class CreationViewModel() : ViewModel() {
     fun getAllRelatedNodes(): MutableList<ComicNode>? {
         return relatedNodes.value?.toMutableList()
     }
+
+    fun setRelatedNodes(items: List<ComicNode>?) {
+        relatedNodes.value = items
+    }
     // endregion
 
     // region: global comic nodes
     fun getGlobalComicNodes() {
         db.collection("globalComicNodes")
-            .orderBy("timeStamp")
+//            .orderBy("timeStamp")
             .limit(50)
             .addSnapshotListener { querySnapshot, ex ->
                 if (ex != null) {
@@ -131,7 +130,7 @@ class CreationViewModel() : ViewModel() {
         }
         db.collection("globalComicNodes")
             .whereEqualTo("ownerUID", FirebaseAuth.getInstance().currentUser!!.uid)
-            .orderBy("timeStamp")
+//            .orderBy("timeStamp")
             .limit(50)
             .addSnapshotListener { querySnapshot, ex ->
                 if (ex != null) {
@@ -163,53 +162,79 @@ class CreationViewModel() : ViewModel() {
     // endregion
 
     // region: data change
-    fun saveComicNode(comicNode: ComicNode) {
+    fun saveComicNode() {
         // inject ids
-        comicNode.selfID = db.collection("globalComicNodes").document().id
-        comicNode.ownerUID = FirebaseAuth.getInstance().currentUser!!.uid
+        if (centerNode.value == null)
+            return
+
+        if (relatedNodes.value != null)
+            centerNode.value!!.relatedNodes = relatedNodes.value
+
+        centerNode.value!!.selfID = db.collection("globalComicNodes").document().id
+        centerNode.value!!.ownerUID = FirebaseAuth.getInstance().currentUser!!.uid
 
         // create document
+        db.collection("globalComicNodes")
+            .document(centerNode.value!!.selfID)
+            .set(centerNode.value!!)
+            .addOnSuccessListener {
+                getGlobalComicNodes()
+                getUserComicNodes()
+            }
+            .addOnFailureListener { e ->
+                Log.d(javaClass.simpleName, e.message)
+            }
+    }
+
+    fun updateComicNode(comicNode: ComicNode) {
         db.collection("globalComicNodes")
             .document(comicNode.selfID)
             .set(comicNode)
             .addOnSuccessListener {
+                getGlobalComicNodes()
+                getUserComicNodes()
             }
             .addOnFailureListener { e ->
+                Log.d(javaClass.simpleName, e.message)
             }
     }
 
-    fun deleteComicNode(comicNode: ComicNode) {
+    fun deleteComicNode(comicNode: ComicNode?) {
+        if (comicNode == null)
+            return
+
         db.collection("globalComicNodes")
             .document(comicNode.selfID)
             .delete()
             .addOnSuccessListener {
             }
             .addOnFailureListener { e ->
+                Log.d(javaClass.simpleName, e.message)
             }
     }
 
     fun moveComicNode(from: Int, to: Int) {
-        if (from == to) return
-        val local = userComicNode.value?.toMutableList()
-        local?.let {
-            // swap order info
-            deleteComicNode(it[from])
-            deleteComicNode(it[to])
-            val temp = it[from].timeStamp
-            it[from].timeStamp = it[to].timeStamp
-            it[to].timeStamp = temp
-            saveComicNode(it[from])
-            saveComicNode(it[to])
-
-            // swap location in list
-            val fromComicNode = it[from]
-            it.removeAt(from)
-            if (to < from) {
-                it.add(to, fromComicNode)
-            } else {
-                it.add(to - 1, fromComicNode)
-            }
-        }
+//        if (from == to) return
+//        val local = userComicNode.value?.toMutableList()
+//        local?.let {
+//            // swap order info
+//            deleteComicNode(it[from])
+//            deleteComicNode(it[to])
+//            val temp = it[from].timeStamp
+//            it[from].timeStamp = it[to].timeStamp
+//            it[to].timeStamp = temp
+//            updateComicNode(it[from])
+//            updateComicNode(it[to])
+//
+//            // swap location in list
+//            val fromComicNode = it[from]
+//            it.removeAt(from)
+//            if (to < from) {
+//                it.add(to, fromComicNode)
+//            } else {
+//                it.add(to - 1, fromComicNode)
+//            }
+//        }
     }
     // endregion
 }
